% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/pda.utils.R
\name{assim.batch}
\alias{assim.batch}
\title{Run Batch PDA}
\usage{
assim.batch(settings)
}
\arguments{
\item{settings}{a PEcAn settings list}
}
\value{
Updated settings list
}
\description{
Run Batch PDA
}
\author{
Ryan Kelly
}

