% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pda.utils.R
\name{correlationPlot}
\alias{correlationPlot}
\title{Flexible function to create correlation density plots
numeric matrix or data.frame}
\usage{
correlationPlot(mat, density = "smooth", thin = "auto",
  method = "pearson", whichParameters = NULL)
}
\arguments{
\item{mat}{matrix or data frame of variables}

\item{density}{type of plot to do}

\item{thin}{thinning of the matrix to make things faster. Default is to thin to 5000}

\item{method}{method for calculating correlations}
}
\description{
Flexible function to create correlation density plots
numeric matrix or data.frame
}
\author{
Florian Hartig
}
\references{
The code for the correlation density plot originates from Hartig, F.; Dislich, C.; Wiegand, T. & Huth, A. (2014) Technical Note: Approximate Bayesian parameterization of a process-based tropical forest model. Biogeosciences, 11, 1261-1272.
}

