% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gelman_diag.R
\name{gelman_diag_gelmanPlot}
\alias{gelman_diag_gelmanPlot}
\title{Calculate Gelman Diagnostic using coda::gelman.plot}
\usage{
gelman_diag_gelmanPlot(x, ...)
}
\arguments{
\item{x}{MCMC samples}

\item{...}{additional arguments}
}
\description{
Calculates Gelman diagnostic cumulatively. This is a much
more conservative approach than the moving-window method.
}
\author{
Alexey Shiklomanov
}
