% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pda.utils.R
\name{generate_hierpost}
\alias{generate_hierpost}
\title{Helper function that generates the hierarchical posteriors}
\usage{
generate_hierpost(mcmc.out, rng_orig)
}
\arguments{
\item{mcmc.out}{hierarchical MCMC outputs}

\item{rng_orig}{nparam x 2 matrix, 1st and 2nd columns are the lower and upper prior limits respectively}
}
\value{
hierarchical MCMC outputs in original parameter space
}
\description{
Helper function that generates the hierarchical posteriors
}
\author{
Istem Fer
}
