% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autoburnin.R
\name{getBurnin}
\alias{getBurnin}
\title{Calculate burnin value}
\usage{
getBurnin(jags_out, threshold = 1.1, use.confidence = TRUE,
  autoburnin = FALSE, plotfile = "/dev/null", ...)
}
\arguments{
\item{jags_out}{List of MCMC sample matrices or `mcmc.list` object}

\item{threshold}{Maximum value of Gelman diagnostic}

\item{use.confidence}{Logical. If TRUE (default), use 95% confidence interval for Gelman Diagnostic.}

\item{autoburnin}{a logical flag indicating whether only the second half
  of the series should be used in the computation.  If set to TRUE
  (default) and \code{start(x)} is less than \code{end(x)/2} then start
  of series will be adjusted so that only second half of series is used.}

\item{plotfile}{Filename for a PNG of the Gelman plot. Default = '/dev/null'
If FALSE, use the point estimate.}

\item{...}{Other parameters to `gelman.plot`}
}
\description{
Use `coda::gelman_plot` to calculate first iteration number where Gelman diagnostic never exceeds the specified threshold.
}
\examples{
     library(coda)
     data(line)
     burnin <- getBurnin(line, threshold = 1.05)
}
\author{
Michael Dietze, Alexey Shiklomanov
}

