% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/minimize.GP.R
\name{gpeval}
\alias{gpeval}
\title{gpeval}
\usage{
gpeval(xnew, k, mu, tau, psi, x, rng, splinefuns)
}
\arguments{
\item{xnew}{new x coordinate}

\item{k}{Specific absorption coefficient (400 - 2500nm)}

\item{mu}{The mean parameter of the distribution; NOTE this is not equal to the mean}

\item{tau}{spatial var}

\item{psi}{spatial corr}

\item{x}{Name of variable to plot on X axis}

\item{rng}{range}

\item{splinefuns}{spline functions}
}
\description{
Calculates the probability of a set of parameter values, given by xnew
}
\author{
Michael Dietze
}
