% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hier.mcmc.R
\name{hier.mcmc}
\alias{hier.mcmc}
\title{Hierarchical MCMC using emulator}
\usage{
hier.mcmc(settings, gp.stack, nstack = NULL, nmcmc, rng_orig, jmp0,
  mu_site_init, nparam, nsites, llik.fn, prior.fn.all, prior.ind.all)
}
\arguments{
\item{settings}{a pecan settings list}

\item{gp.stack}{list of GPs}

\item{nstack}{list of number of observations, currently not used}

\item{nmcmc}{number of MCMC iterations}

\item{rng_orig}{range of knots}

\item{jmp0}{initial jump vars}

\item{mu_site_init}{initial parameter values per site}

\item{nparam}{number of parameters}

\item{nsites}{number of sites}

\item{llik.fn}{list of likelihood functions}

\item{prior.fn.all}{list of prior functions}

\item{prior.ind.all}{indices of targeted parameters}
}
\description{
Hierarchical MCMC
}
\author{
Istem Fer
}
