% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load.L2Ameriflux.cf.R, R/pda.load.data.R
\name{load.pda.data}
\alias{load.pda.data}
\alias{load.L2Ameriflux.cf}
\title{Load Ameriflux L2 Data From NetCDF}
\usage{
load.L2Ameriflux.cf(file.in)

load.pda.data(settings, bety)
}
\arguments{
\item{file.in}{= the netcdf file of L2 data}

\item{settings}{= PEcAn settings list}
}
\value{
A data frame of all variables in the netcdf

A list containg the loaded input data, plus metadata
}
\description{
Load Ameriflux L2 Data From NetCDF

Load Dataset for Paramater Data Assimilation
}
\author{
Ryan Kelly

Ryan Kelly, Istem Fer
}
