% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pda.utils.R
\name{load_pda_history}
\alias{load_pda_history}
\title{Helper function that loads history from previous PDA run, but returns only requested objects}
\usage{
load_pda_history(workdir, ensemble.id, objects)
}
\description{
Helper function that loads history from previous PDA run, but returns only requested objects
}
\author{
Istem Fer
}
