% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autoburnin.R
\name{makeMCMCList}
\alias{makeMCMCList}
\title{Make MCMC list from samples list}
\usage{
makeMCMCList(samps)
}
\arguments{
\item{samps}{samples list (output from invert.custom)}
}
\description{
Make MCMC list from samples list
}
