% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/minimize.GP.R
\name{minimize.GP}
\alias{minimize.GP}
\title{minimize.GP}
\usage{
minimize.GP(gp, rng, x0, splinefuns = NULL)
}
\arguments{
\item{gp}{Gaussian Process}

\item{rng}{range}

\item{x0}{initial values}

\item{splinefuns}{spline functions}
}
\description{
minimize.GP
}
\author{
Michael Dietze
}
