% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pda.utils.R
\name{norm_transform_priors}
\alias{norm_transform_priors}
\title{Helper function that transforms the values of each parameter into N(0,1) equivalent}
\usage{
norm_transform_priors(prior.list, prior.fn.all, prior.ind.all, SS.stack,
  init.list, jmp.list)
}
\description{
Helper function that transforms the values of each parameter into N(0,1) equivalent
}
\author{
Istem Fer
}
