% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pda.utils.R
\name{pda.adjust.jumps.bs}
\alias{pda.adjust.jumps.bs}
\title{Adjust PDA block MCMC jump size}
\usage{
pda.adjust.jumps.bs(settings, jcov, accept.count, params.recent)
}
\arguments{
\item{settings}{a PEcAn settings list}

\item{jcov}{jump covariance matrix}

\item{accept.count}{aceeptance count}

\item{params.recent}{parameters accepted since previous adjustment}
}
\value{
A PEcAn settings list updated to reflect adjusted jump distributions
}
\description{
Adjust PDA blcok MCMC jump size
}
\author{
Ryan Kelly
}
