% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pda.utils.R
\name{pda.adjust.jumps}
\alias{pda.adjust.jumps}
\title{Adjust PDA MCMC jump size}
\usage{
pda.adjust.jumps(settings, jmp.list, accept.rate, pnames = NULL)
}
\arguments{
\item{settings}{a PEcAn settings list}

\item{jmp.list}{list of jump variances}

\item{accept.rate}{acceptance rate}

\item{pnames}{parameter names}
}
\value{
A PEcAn settings list updated to reflect adjusted jump distributions
}
\description{
Adjust PDA MCMC jump size
}
\author{
Ryan Kelly
}
