% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pda.neff.R
\name{pda.autocorr.calc}
\alias{pda.autocorr.calc}
\title{autocorrelation correction}
\usage{
pda.autocorr.calc(input, model = "heteroskedastic.laplacian")
}
\arguments{
\item{input}{list that contains time-series data vector and parameters for heteroskedastic.laplacian}

\item{model}{data model type, for flux data heteroskedastic laplacian, normal is an example}
}
\value{
rho AR(1)
}
\description{
autocorrelation correction
}
\author{
Istem Fer
}
