% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pda.define.llik.R
\name{pda.calc.llik.par}
\alias{pda.calc.llik.par}
\title{pda.calc.llik.par}
\usage{
pda.calc.llik.par(settings, n, error.stats, hyper.pars)
}
\arguments{
\item{settings}{list}

\item{n}{named vector, sample sizes of inputs}

\item{error.stats}{list, Sufficient Statistics}

\item{hyper.pars}{list, hyperparameters}
}
\description{
Calculate likelihood parameters
}
\author{
Istem Fer
}
