% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pda.define.llik.R
\name{pda.calc.llik}
\alias{pda.calc.llik}
\title{Calculate Likelihoods for PDA}
\usage{
pda.calc.llik(pda.errors, llik.fn, llik.par)
}
\arguments{
\item{pda.errors}{calculated errors}

\item{llik.fn}{list of likelihood fcns}

\item{llik.par}{parameters to be passed llik functions}
}
\value{
Total log likelihood (i.e., sum of log likelihoods for each dataset)
}
\description{
Calculate Likelihoods for PDA
}
\author{
Ryan Kelly, Istem Fer
}
