% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pda.utils.R
\name{pda.create.ensemble}
\alias{pda.create.ensemble}
\title{Create ensemble record for PDA ensemble}
\usage{
pda.create.ensemble(settings, con, workflow.id)
}
\arguments{
\item{settings}{a PEcAn settings list}

\item{con}{DB connection}

\item{workflow.id}{workflow ID}
}
\value{
Ensemble ID of the created ensemble
}
\description{
Create PDA Ensemble
}
\author{
Ryan Kelly
}
