% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pda.define.llik.R
\name{pda.define.llik.fn}
\alias{pda.define.llik.fn}
\title{Define PDA Likelihood Functions}
\usage{
pda.define.llik.fn(settings)
}
\arguments{
\item{settings}{PEcAn settings list}
}
\value{
List of likelihood functions, one for each dataset to be assimilated against.
}
\description{
Define PDA Likelihood Functions
}
\author{
Ryan Kelly, Istem Fer
}
