% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pda.emulator.ms.R
\name{pda.emulator.ms}
\alias{pda.emulator.ms}
\title{Paramater Data Assimilation using emulator on multiple sites}
\usage{
pda.emulator.ms(settings, params.id = NULL, param.names = NULL,
  prior.id = NULL, chain = NULL, iter = NULL, adapt = NULL,
  adj.min = NULL, ar.target = NULL, jvar = NULL, n.knot = NULL)
}
\arguments{
\item{settings}{= a pecan settings list}
}
\value{
settings
}
\description{
Paramater Data Assimilation using emulator on multiple sites in three modes: local, global, hierarchical
}
\author{
Istem Fer
}
