% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pda.utils.R
\name{pda.generate.knots}
\alias{pda.generate.knots}
\title{Generate Parameter Knots for PDA Emulator}
\usage{
pda.generate.knots(n.knot, sf, probs.sf, n.param.all, prior.ind, prior.fn,
  pname)
}
\arguments{
\item{n.knot}{number of knots}

\item{sf}{scaling factor}

\item{probs.sf}{values for sf}

\item{n.param.all}{number of all params}

\item{prior.ind}{indices of targeted parameters in the prior dataframe}

\item{prior.fn}{list of prior functions}

\item{pname}{name of parameters}
}
\value{
A list of probabilities and parameter values, with one row for each knot in the emulator.
}
\description{
Generate Parameter Knots for PDA Emulator
}
\author{
Ryan Kelly, Istem Fer
}
