% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pda.utils.R
\name{pda.generate.sf}
\alias{pda.generate.sf}
\title{Generate scaling factor knots for PDA Emulator}
\usage{
pda.generate.sf(n.knot, sf, prior.list)
}
\arguments{
\item{n.knot}{number of knots}

\item{sf}{scaling factor}

\item{prior.list}{list of prior dataframes}
}
\description{
Generate scaling factor knots for PDA Emulator
}
\author{
Istem Fer
}
