% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pda.get.model.output.R
\name{pda.get.model.output}
\alias{pda.get.model.output}
\title{Get Model Output for PDA}
\usage{
pda.get.model.output(settings, run.id, bety, inputs, external.formats = NULL)
}
\arguments{
\item{settings}{PEcAn settings list}

\item{run.id}{run ID}

\item{bety}{bety list}

\item{inputs}{inputs list}

\item{external.formats}{format list}
}
\value{
A list containing model outputs extracted to correspond to each observational
dataset being used for PDA.
}
\description{
Get Model Output for PDA
}
\author{
Ryan Kelly, Istem Fer
}
