% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pda.utils.R
\name{pda.init.params}
\alias{pda.init.params}
\title{Initialise Parameter Matrix for PDA}
\usage{
pda.init.params(settings, chain, pname, n.param.all)
}
\arguments{
\item{settings}{a PEcAn settings list}

\item{chain}{number of chain}

\item{pname}{parameter name}

\item{n.param.all}{number of all params}
}
\value{
A list containing 'start' and 'finish' counters for MCMC, as well as the params
table, which is an empty matrix concatenated to any param samples from a previous
PDA run, if provided.
}
\description{
Initialise Parameter Matrix for PDA
}
\author{
Ryan Kelly
}
