% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pda.utils.R
\name{pda.init.run}
\alias{pda.init.run}
\title{Initialise Model Runs for PDA}
\usage{
pda.init.run(settings, con, my.write.config, workflow.id, params,
  n = ifelse(is.null(dim(params)), 1, nrow(params)),
  run.names = paste("run", 1:n, sep = "."))
}
\arguments{
\item{settings}{a PEcAn settings list}

\item{con}{DB connection}

\item{my.write.config}{model write config fcn name}

\item{workflow.id}{workflow ID}

\item{params}{parameters of the run}

\item{n}{number of runs}

\item{run.names}{names of runs}
}
\value{
Vector of run IDs for all model runs that were set up (including write.configs)
}
\description{
Initialise Model Runs for PDA
}
\author{
Ryan Kelly
}
