% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pda.utils.R
\name{pda.load.priors}
\alias{pda.load.priors}
\title{Load Priors for Paramater Data Assimilation}
\usage{
pda.load.priors(settings, con, extension.check = FALSE)
}
\arguments{
\item{settings}{a PEcAn settings list}

\item{con}{database connection}

\item{extension.check}{check if this is another round or longer run}
}
\value{
A previously-generated posterior distribution, to be used as the prior for PDA.
}
\description{
Load Priors for Paramater Data Assimilation
}
\author{
Ryan Kelly, Istem Fer
}
