% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pda.neff.R
\name{pda.neff.calc}
\alias{pda.neff.calc}
\title{Calculate N_eff}
\usage{
pda.neff.calc(inputs)
}
\arguments{
\item{inputs}{list}
}
\value{
inputs list, updated inputs with n_eff
}
\description{
Autocorelation correction and efficient sample size calculation on latent process
}
\author{
Istem Fer
}

