% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pda.postprocess.R
\name{pda.plot.params}
\alias{pda.plot.params}
\title{Plot PDA Parameter Diagnostics}
\usage{
pda.plot.params(
  settings,
  mcmc.param.list,
  prior.ind,
  par.file.name = NULL,
  sffx
)
}
\arguments{
\item{settings}{PEcAn settings list}

\item{mcmc.param.list}{MCMC param list to be sorted}

\item{prior.ind}{indices of the targeted parameters}

\item{par.file.name}{output file name}

\item{sffx}{suffix to the output file names}
}
\value{
Nothing. Plot is generated and saved to PDF.
}
\description{
Plot PDA Parameter Diagnostics
}
\author{
Ryan Kelly, Istem Fer
}
