% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pda.postprocess.R
\name{pda.postprocess}
\alias{pda.postprocess}
\title{Postprocessing for PDA Results}
\usage{
pda.postprocess(settings, con, mcmc.param.list, pname, prior, prior.ind,
  sffx = NULL)
}
\arguments{
\item{settings}{PEcAn settings list}

\item{con}{DB connection}

\item{mcmc.param.list}{output of PDA MCMC}

\item{pname}{parameter names}

\item{prior}{prior list}

\item{prior.ind}{indices of targeted parameters}

\item{sffx}{suffix to the output files, e.g. "hierarchical"}
}
\value{
PEcAn settings list, updated with <params.id> pointing to the new params file.
}
\description{
Postprocessing for PDA Results
}
\author{
Ryan Kelly, Istem Fer
}
