% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pda.bayestools.helpers.R
\name{pda.settings.bt}
\alias{pda.settings.bt}
\title{Apply settings for BayesianTools}
\usage{
pda.settings.bt(settings)
}
\arguments{
\item{settings}{PEcAn settings}
}
\value{
bt.settings list of runMCMC{BayesianTools} settings
}
\description{
Helper function for applying BayesianTools specific settings from PEcAn general settings
}
\author{
Istem Fer
}
