% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pda.utils.R
\name{pda.settings}
\alias{pda.settings}
\title{Set PDA Settings}
\usage{
pda.settings(
  settings,
  params.id = NULL,
  param.names = NULL,
  prior.id = NULL,
  chain = NULL,
  iter = NULL,
  adapt = NULL,
  adj.min = NULL,
  ar.target = NULL,
  jvar = NULL,
  n.knot = NULL,
  run.round = FALSE
)
}
\arguments{
\item{settings}{a PEcAn settings list}

\item{params.id}{id of pars}

\item{param.names}{names of pars}

\item{prior.id}{ids of priors}

\item{chain}{how many chains}

\item{iter}{how many iterations}

\item{adapt}{adaptation intervals}

\item{adj.min}{to be used in adjustment}

\item{ar.target}{acceptance rate target}

\item{jvar}{jump variance}

\item{n.knot}{number of knots requested}

\item{run.round}{another round or not}
}
\value{
An updated settings list
}
\description{
Set PDA Settings
}
\author{
Ryan Kelly, Istem Fer
}
