% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pda.postprocess.R
\name{pda.sort.params}
\alias{pda.sort.params}
\title{Function to sort Hierarchical MCMC samples}
\usage{
pda.sort.params(mcmc.out, sub.sample = "mu_global_samp", ns = NULL,
  prior.all, prior.ind.all.ns, sf = NULL, n.param.orig, prior.list,
  prior.fn.all)
}
\description{
Function to sort Hierarchical MCMC samples
}
