% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pda.utils.R
\name{prepare_pda_remote}
\alias{prepare_pda_remote}
\title{helper function for submitting remote pda runs}
\usage{
prepare_pda_remote(settings, site = 1, multi_site_objects)
}
\arguments{
\item{settings}{PEcAn settings list}

\item{site}{site number (which site)}

\item{multi_site_objects}{information needed for remote runs}
}
\description{
helper function for submitting remote pda runs
}
