% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pda.utils.R
\name{return_hyperpars}
\alias{return_hyperpars}
\title{return_hyperpars}
\usage{
return_hyperpars(assim.settings, inputs)
}
\arguments{
\item{assim.settings}{PEcAn settings list}

\item{inputs}{inputs list}
}
\description{
return_hyperpars
}
\author{
Istem Fer
}
