% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pda.utils.R
\name{return_multi_site_objects}
\alias{return_multi_site_objects}
\title{This is a helper function partly uses pda.emulator code}
\usage{
return_multi_site_objects(multi.settings)
}
\arguments{
\item{multi.settings}{PEcAn multi settings object}
}
\description{
This is a helper function partly uses pda.emulator code
}
