% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pda.utils.R
\name{runModule.assim.batch}
\alias{runModule.assim.batch}
\title{Run Batch module}
\usage{
runModule.assim.batch(settings)
}
\arguments{
\item{settings}{a PEcAn settings list}
}
\description{
Run Batch module
}
