% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pda.utils.R
\name{sample_MCMC}
\alias{sample_MCMC}
\title{Helper function to sample from previous MCMC chain while proposing new knots}
\usage{
sample_MCMC(mcmc_path, n.param.orig, prior.ind.orig, n.post.knots,
  knots.params.temp)
}
\description{
Helper function to sample from previous MCMC chain while proposing new knots
}
\author{
Istem Fer
}
