% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pda.utils.R
\name{sync_pda_remote}
\alias{sync_pda_remote}
\title{helper function for syncing remote pda runs
this function resembles remote.copy.from but we don't want to sync everything back}
\usage{
sync_pda_remote(multi.settings, ensembleidlist, register = FALSE)
}
\arguments{
\item{multi.settings}{PEcAn multi settings}

\item{ensembleidlist}{ensemble id list for remote runs}

\item{register}{if register==TRUE, the last files returned will be registered to the DB, TO BE DONE}
}
\description{
helper function for syncing remote pda runs
this function resembles remote.copy.from but we don't want to sync everything back
}
