% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pda.postprocess.R
\name{write_sf_posterior}
\alias{write_sf_posterior}
\title{Function to write posterior distributions of the scaling factors}
\usage{
write_sf_posterior(sf.samp.list, sf.prior, sf.samp.filename)
}
\description{
Function to write posterior distributions of the scaling factors
}
