% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autoburnin.R
\name{autoburnin}
\alias{autoburnin}
\title{Automatically calculate and apply burnin value}
\usage{
autoburnin(jags_out, ...)
}
\arguments{
\item{jags_out}{List of MCMC sample matrices or `mcmc.list` object}
}
\author{
Michael Dietze, Alexey Shiklomanov
}

