% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autoburnin.R
\name{autoburnin}
\alias{autoburnin}
\title{Automatically calculate and apply burnin value}
\usage{
autoburnin(jags_out, return.burnin = FALSE, ...)
}
\arguments{
\item{jags_out}{List of MCMC sample matrices or `mcmc.list` object}

\item{return.burnin}{Logical. If `TRUE`, return burnin value in addition to 
samples (as list). Default = FALSE.}

\item{...}{Additional arguments for \code{getBurnin}, \code{gelman_diag_mw}, 
and \code{gelman.diag}.}
}
\description{
Automatically calculate and apply burnin value
}
\examples{
     z1 <- coda::mcmc(c(rnorm(2500, 5), rnorm(2500, 0)))
     z2 <- coda::mcmc(c(rnorm(2500, -5), rnorm(2500, 0)))
     z <- coda::mcmc.list(z1, z2)
     z_burned <- autoburnin(z)
}
\author{
Michael Dietze, Alexey Shiklomanov
}
