% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pda.utils.R
\name{back_transform_posteriors}
\alias{back_transform_posteriors}
\title{Helper function that transforms the samples back to their original prior distribution equivalents}
\usage{
back_transform_posteriors(prior.list, prior.fn.all, prior.ind.all, mcmc.out)
}
\description{
Helper function that transforms the samples back to their original prior distribution equivalents
}
\author{
Istem Fer
}
