% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gelman_diag.R
\name{gelman_diag_mw}
\alias{gelman_diag_mw}
\title{Calculate Gelman diagnostic on moving window}
\usage{
gelman_diag_mw(x, width_fraction = 0.1, width = ceiling(coda::niter(x)
  * width_fraction), njump = 50, include.mpsrf = TRUE, ...)
}
\arguments{
\item{x}{MCMC samples, of class \code{mcmc} or \code{mcmc.list}}

\item{width_fraction}{Fractional width of moving window. Default=0.1.}

\item{width}{Width of moving window. Default is niter(x)*width_fraction}

\item{njump}{Number of windows to calculate over}

\item{include.mpsrf}{Whether to calculate multivariate PSRF and include in output (default = FALSE).}
}
\value{
Gelman Diagnostic 3D array. First dim -- mean (1) and 95% confidence (2). Second dim -- iteration
}
\description{
Calculate Gelman diagnostic on moving window
}
\author{
Alexey Shiklomanov
}
