% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pda.utils.R
\name{generate_hierpost}
\alias{generate_hierpost}
\title{Helper function that generates the hierarchical posteriors}
\usage{
generate_hierpost(mcmc.out, prior.fn.all, prior.ind.all)
}
\arguments{
\item{mcmc.out}{hierarchical MCMC outputs}

\item{prior.fn.all}{list of all prior functions}

\item{prior.ind.all}{indices of the targeted params}
}
\value{
hierarchical MCMC outputs in original parameter space
}
\description{
Helper function that generates the hierarchical posteriors
}
\author{
Istem Fer
}
