% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autoburnin.R
\name{getBurnin}
\alias{getBurnin}
\title{Calculate burnin value}
\usage{
getBurnin(jags_out, threshold = 1.1, use.confidence = TRUE,
  method = "moving.window", plotfile = "/dev/null", ...)
}
\arguments{
\item{jags_out}{List of MCMC sample matrices or `mcmc.list` object}

\item{threshold}{Maximum value of Gelman diagnostic}

\item{use.confidence}{Logical. If TRUE (default), use 95% confidence 
interval for Gelman Diagnostic.
If FALSE, use the point estimate.}

\item{method}{Character string indicating method (default="moving.window". 
No other options are currently implemented.)}

\item{...}{Other parameters to methods}

\item{width}{Width of window for "moving.window" method. Default = niter / 2}

\item{njump}{Number of steps for "moving.window" method. Default = 50}
}
\description{
Automatically detect burnin based on one of several methods.
}
\details{
See "gelman_diag_mw".
}
\examples{
     z1 <- coda::mcmc(c(rnorm(2500, 5), rnorm(2500, 0)))
     z2 <- coda::mcmc(c(rnorm(2500, -5), rnorm(2500, 0)))
     z <- coda::mcmc.list(z1, z2)
     burnin <- getBurnin(z, threshold = 1.05)
}
\author{
Alexey Shiklomanov, Michael Dietze
}

