% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autoburnin.R
\name{getBurnin}
\alias{getBurnin}
\title{Calculate burnin value}
\usage{
getBurnin(jags_out, threshold = 1.1, ...)
}
\arguments{
\item{jags_out}{List of MCMC sample matrices or `mcmc.list` object}

\item{threshold}{Maximum value of Gelman diagnostic}

\item{...}{Other parameters to `gelman.plot`}
}
\description{
Use `coda::gelman_plot` to calculate first iteration number where Gelman diagnostic never exceeds the specified threshold.
}
\examples{
     library(coda)
     data(line)
     burnin <- getBurnin(line, threshold = 1.05)
}
\author{
Michael Dietze, Alexey Shiklomanov
}

