% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autoburnin.R
\name{getBurnin}
\alias{getBurnin}
\title{Calculate burnin value}
\usage{
getBurnin(jags_out, method = "rmw", threshold = 1.1,
  use.confidence = TRUE, width = ceiling(coda::niter(jags_out)/2),
  njump = 50, autoburnin = FALSE, plotfile = "/dev/null", ...)
}
\arguments{
\item{jags_out}{List of MCMC sample matrices or `mcmc.list` object}

\item{method}{Character string indicating method ("rmw" or "gelman.plot")}

\item{threshold}{Maximum value of Gelman diagnostic}

\item{use.confidence}{Logical. If TRUE (default), use 95% confidence interval for Gelman Diagnostic.
If FALSE, use the point estimate.}

\item{width}{Width of window for `rmw` method. Default = niter / 2}

\item{njump}{Number of steps for `rmw` method. Default = 50}

\item{autoburnin}{a logical flag indicating whether only the second half
  of the series should be used in the computation.  If set to TRUE
  (default) and \code{start(x)} is less than \code{end(x)/2} then start
  of series will be adjusted so that only second half of series is used.}

\item{plotfile}{Filename for a PNG of the Gelman plot. Default = '/dev/null'}

\item{...}{Other parameters to `gelman.plot`}
}
\description{
Automatically detect burnin based on one of several methods.
}
\details{
rmw (Reverse moving window) method divides the chains into `njump` sections, each of size `width`, and loops over them starting from the last one until the first block that has not converged.
The iteration starting the next block after that is the burnin.

`gelman.plot` method uses the `gelman.plot` function from the `coda` package to calculate the Gelman diagnostic cumulatively from the beginning of the chain.
The iteration after the last time the `gelman.diag` exceeds the threshold for any of the parameters is the burnin.
}
\examples{
     library(coda)
     data(line)
     burnin <- getBurnin(line, threshold = 1.05, 
                         method = 'gelman.plot')
}
\author{
Alexey Shiklomanov, Michael Dietze
}

