% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autoburnin.R
\name{getBurnin}
\alias{getBurnin}
\title{Calculate burnin value}
\usage{
getBurnin(jags_out, threshold = 1.1, use.confidence = TRUE,
  method = "moving.window", plotfile = "/dev/null", ...)
}
\arguments{
\item{jags_out}{List of MCMC sample matrices or \code{mcmc.list} object}

\item{threshold}{Maximum value of Gelman diagnostic}

\item{use.confidence}{Logical. If TRUE (default), use 95% confidence
interval for Gelman Diagnostic. If FALSE, use the point estimate.}

\item{method}{Character string indicating method. Options are
"moving.window" (default) or "gelman.plot".}

\item{plotfile}{path}

\item{...}{Other parameters to methods}
}
\description{
Automatically detect burnin based on one of several methods.
}
\details{
See "gelman_diag_mw" and "gelman_diag_gelmanPlot"
}
\examples{
     z1 <- coda::mcmc(c(rnorm(2500, 5), rnorm(2500, 0)))
     z2 <- coda::mcmc(c(rnorm(2500, -5), rnorm(2500, 0)))
     z <- coda::mcmc.list(z1, z2)
     burnin <- getBurnin(z, threshold = 1.05)
}
\author{
Alexey Shiklomanov, Michael Dietze
}
