% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/load.L2Ameriflux.cf.R, R/pda.utils.R
\name{load.pda.data}
\alias{load.L2Ameriflux.cf}
\alias{load.pda.data}
\title{Load Ameriflux L2 Data From NetCDF}
\usage{
load.L2Ameriflux.cf(file.in)

load.pda.data(input.settings, con)
}
\arguments{
\item{file.in}{= the netcdf file of L2 data}

\item{input.settings}{= settings$assim.batch$inputs from pecan.xml or similar}
}
\value{
A data frame of all variables in the netcdf

A list containg the loaded input data, plus metadata
}
\description{
Load Ameriflux L2 Data From NetCDF

Load Dataset for Paramater Data Assimilation
}
\author{
Ryan Kelly

Ryan Kelly
}

