% Generated by roxygen2 (4.0.2): do not edit by hand
\name{load.pda.data}
\alias{load.L2Ameriflux.cf}
\alias{load.pda.data}
\title{Load Ameriflux L2 Data From NetCDF}
\usage{
load.L2Ameriflux.cf(file.in)

load.pda.data(settings, con)
}
\arguments{
\item{file.in}{= the netcdf file of L2 data}

\item{settings}{= PEcAn settings list}
}
\value{
A data frame of all variables in the netcdf

A list containg the loaded input data, plus metadata
}
\description{
Load Ameriflux L2 Data From NetCDF

Load Dataset for Paramater Data Assimilation
}
\author{
Ryan Kelly

Ryan Kelly
}

