% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/pda.utils.R
\name{pda.adjust.jumps}
\alias{pda.adjust.jumps}
\title{Adjust PDA MCMC jump size}
\usage{
pda.adjust.jumps(settings, accept.rate, pnames = NULL)
}
\arguments{
\item{all}{params are the identically named variables in pda.mcmc / pda.emulator}
}
\value{
A PEcAn settings list updated to reflect adjusted jump distributions
}
\description{
Adjust PDA MCMC jump size
}
\author{
Ryan Kelly
}

