% Generated by roxygen2 (4.0.2): do not edit by hand
\name{pda.bayesian.tools}
\alias{pda.bayesian.tools}
\title{Paramater Data Assimilation using BayesianTools}
\usage{
pda.bayesian.tools(settings, params.id = NULL, param.names = NULL,
  prior.id = NULL, chain = NULL, iter = NULL, adapt = NULL,
  adj.min = NULL, ar.target = NULL, jvar = NULL, n.knot = NULL)
}
\arguments{
\item{settings}{= a pecan settings list}
}
\value{
nothing. Diagnostic plots, MCMC samples, and posterior distributions
 are saved as files and db records.
}
\description{
Paramater Data Assimilation using BayesianTools R Package
}
\author{
Mike Dietze

Istem Fer, Ryan Kelly
}

