% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pda.define.llik.R
\name{pda.calc.error}
\alias{pda.calc.error}
\title{Calculate Likelihood parameters}
\usage{
pda.calc.error(settings, con, model_out, run.id, inputs)
}
\arguments{
\item{settings}{list}

\item{inputs}{list}

\item{model.out}{list}
}
\value{
inputs updated inputs list with likelihood parameters
}
\description{
Calculate Likelihood parameters
}
\author{
Istem Fer
}

