% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pda.define.llik.R
\name{pda.calc.error}
\alias{pda.calc.error}
\title{Calculate sufficient statistics}
\usage{
pda.calc.error(settings, con, model_out, run.id, inputs, bias.terms)
}
\arguments{
\item{settings}{list}

\item{con}{DB connection}

\item{model_out}{list}

\item{run.id}{run ID}

\item{inputs}{list}

\item{bias.terms}{matrix}
}
\value{
pda.errors
}
\description{
Calculate sufficient statistics
}
\author{
Istem Fer
}
