% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/pda.utils.R
\name{pda.calc.llik}
\alias{pda.calc.llik}
\title{Calculate Likelihoods for PDA}
\usage{
pda.calc.llik(settings, con, model.out, run.id, inputs, llik.fn)
}
\arguments{
\item{all}{params are the identically named variables in pda.mcmc / pda.emulator}
}
\value{
Total log likelihood (i.e., sum of log likelihoods for each dataset)
}
\description{
Calculate Likelihoods for PDA
}
\author{
Ryan Kelly
}

