% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pda.bayestools.helpers.R
\name{pda.create.btprior}
\alias{pda.create.btprior}
\title{Create priors for BayesianTools}
\usage{
pda.create.btprior(prior.sel)
}
\arguments{
\item{prior.sel}{`data.frame` containing prior distributions of the selected parameters}
}
\value{
out Prior class object for BayesianTools package
}
\description{
Helper function for creating log-priors compatible with BayesianTools package
}
\details{
`prior.sel` must contain the following columns:
  * `distn` -- String describing a distribution; e.g. `norm` for `dnorm`, `rnorm`, etc.
  * `parama`, `paramb` -- First and second parameters, respectively, of the corresponding distribution

Optionally, `prior.sel` may also contain the following columns:
  * `param_name` -- Parameter name, which will be carried through to the prior object and sampler
  * `lower`, `upper` -- Lower and upper bounds, respectively. These can be leveraged by the BayesianTools samplers.
  * `best` -- Best guess for a parameter estimate. BayesianTools can also use this, though I'm not sure how...
}
\author{
Istem Fer, Alexey Shiklomanov
}
