% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pda.bayestools.helpers.R
\name{pda.create.btprior}
\alias{pda.create.btprior}
\title{Create priors for BayesianTools}
\usage{
pda.create.btprior(prior.sel)
}
\arguments{
\item{prior.sel}{\code{data.frame} containing prior distributions of the selected parameters}
}
\value{
out Prior class object for BayesianTools package
}
\description{
Helper function for creating log-priors compatible with BayesianTools package
}
\details{
\code{prior.sel} must contain the following columns:
\itemize{
\item \code{distn} -- String describing a distribution; e.g. \code{norm} for \code{dnorm}, \code{rnorm}, etc.
\item \code{parama}, \code{paramb} -- First and second parameters, respectively, of the corresponding distribution
}

Optionally, \code{prior.sel} may also contain the following columns:
\itemize{
\item \code{param_name} -- Parameter name, which will be carried through to the prior object and sampler
\item \code{lower}, \code{upper} -- Lower and upper bounds, respectively. These can be leveraged by the BayesianTools samplers.
\item \code{best} -- Best guess for a parameter estimate. BayesianTools can also use this, though I'm not sure how...
}
}
\author{
Istem Fer, Alexey Shiklomanov
}
