% Generated by roxygen2 (4.0.2): do not edit by hand
\name{pda.create.btprior}
\alias{pda.create.btprior}
\title{Create priors for BayesianTools}
\usage{
pda.create.btprior(prior.sel)
}
\arguments{
\item{prior.sel}{prior distributions of the selected parameters}
}
\value{
out prior class object for BayesianTools package
}
\description{
Helper function for creating log-priors compatible with BayesianTools package
}
\author{
Istem Fer
}

