% Generated by roxygen2 (4.0.2): do not edit by hand
\name{pda.create.ensemble}
\alias{pda.create.ensemble}
\title{Create ensemble record for PDA ensemble}
\usage{
pda.create.ensemble(settings, con, workflow.id)
}
\arguments{
\item{all}{params are the identically named variables in pda.mcmc / pda.emulator}
}
\value{
Ensemble ID of the created ensemble
}
\description{
Create PDA Ensemble
}
\author{
Ryan Kelly
}

