% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/pda.utils.R
\name{pda.define.llik.fn}
\alias{pda.define.llik.fn}
\title{Define PDA Likelihood Functions}
\usage{
pda.define.llik.fn(settings)
}
\arguments{
\item{all}{params are the identically named variables in pda.mcmc / pda.emulator}
}
\value{
List of likelihood functions, one for each dataset to be assimilated against.
}
\description{
Define PDA Likelihood Functions
}
\author{
Ryan Kelly
}

