% Generated by roxygen2 (4.0.2): do not edit by hand
\name{pda.define.prior.fn}
\alias{pda.define.prior.fn}
\title{Define PDA Prior Functions}
\usage{
pda.define.prior.fn(prior)
}
\arguments{
\item{all}{params are the identically named variables in pda.mcmc / pda.emulator}
}
\value{
List of prior functions containing dprior, rprior, qprior, dmvprior, rmvprior.
        Each of these is a list with one distribution function per parameter.
}
\description{
Define PDA Prior Functions
}
\author{
Ryan Kelly
}

