% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pda.emulator.ms.R
\name{pda.emulator.ms}
\alias{pda.emulator.ms}
\title{Paramater Data Assimilation using emulator on multiple sites in three modes: local, global, hierarchical
First draft, not complete yet}
\usage{
pda.emulator.ms(multi.settings)
}
\arguments{
\item{multi.settings}{= a pecan multi-settings list}
}
\value{
settings
}
\description{
Paramater Data Assimilation using emulator on multiple sites in three modes: local, global, hierarchical
First draft, not complete yet
}
\author{
Istem Fer
}
