% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pda.emulator.R
\name{pda.emulator}
\alias{pda.emulator}
\title{Paramater Data Assimilation using emulator}
\usage{
pda.emulator(settings, external.data = NULL, external.priors = NULL,
  params.id = NULL, param.names = NULL, prior.id = NULL,
  chain = NULL, iter = NULL, adapt = NULL, adj.min = NULL,
  ar.target = NULL, jvar = NULL, n.knot = NULL)
}
\arguments{
\item{settings}{= a pecan settings list}

\item{external.data}{= list of inputs}

\item{external.priors}{= list or priors}
}
\value{
nothing. Diagnostic plots, MCMC samples, and posterior distributions
 are saved as files and db records.
}
\description{
Paramater Data Assimilation using emulator
}
\author{
Mike Dietze

Ryan Kelly, Istem Fer
}
