% Generated by roxygen2 (4.0.2): do not edit by hand
\name{pda.generate.knots}
\alias{pda.generate.knots}
\title{Generate Parameter Knots for PDA Emulator}
\usage{
pda.generate.knots(n.knot, n.param.all, prior.ind, prior.fn, pname)
}
\arguments{
\item{all}{params are the identically named variables in pda.mcmc / pda.emulator}
}
\value{
A matrix of parameter values, with one row for each knot in the emulator.
}
\description{
Generate Parameter Knots for PDA Emulator
}
\author{
Ryan Kelly
}

