% Generated by roxygen2 (4.0.2): do not edit by hand
\name{pda.get.model.output}
\alias{pda.get.model.output}
\title{Get Model Output for PDA}
\usage{
pda.get.model.output(settings, run.id, inputs)
}
\arguments{
\item{all}{params are the identically named variables in pda.mcmc / pda.emulator}
}
\value{
A list containing model outputs extracted to correspond to each observational
        dataset being used for PDA.
}
\description{
Get Model Output for PDA
}
\author{
Ryan Kelly
}

