% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/pda.utils.R
\name{pda.init.run}
\alias{pda.init.run}
\title{Initialise Model Runs for PDA}
\usage{
pda.init.run(settings, con, my.write.config, workflow.id, params,
  n = ifelse(is.null(dim(params)), 1, nrow(params)),
  run.names = paste("run", 1:n, sep = "."))
}
\arguments{
\item{all}{params are the identically named variables in pda.mcmc / pda.emulator}
}
\value{
Vector of run IDs for all model runs that were set up (including write.configs)
}
\description{
Initialise Model Runs for PDA
}
\author{
Ryan Kelly
}

