% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pda.utils.R
\name{pda.load.priors}
\alias{pda.load.priors}
\title{Load Priors for Paramater Data Assimilation}
\usage{
pda.load.priors(settings, con)
}
\arguments{
\item{all}{params are the identically named variables in pda.mcmc / pda.emulator}
}
\value{
A previously-generated posterior distribution, to be used as the prior for PDA.
}
\description{
Load Priors for Paramater Data Assimilation
}
\author{
Ryan Kelly
}

