% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/pda.mcmc.R
\name{pda.mcmc}
\alias{pda.mcmc}
\title{Paramater Data Assimilation using MCMC}
\usage{
pda.mcmc(settings)
}
\arguments{
\item{settings}{= a pecan settings list}
}
\value{
nothing. Diagnostic plots, MCMC samples, and posterior distributions
 are saved as files and db records.
}
\description{
Paramater Data Assimilation using MCMC
}
\details{
Brute-force, only to be used on simple models
}
\author{
Mike Dietze

Ryan Kelly
}

