% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/pda.mcmc.recover.R
\name{pda.mcmc.recover}
\alias{pda.mcmc.recover}
\title{Clean up a failed PDA run}
\usage{
pda.mcmc.recover(settings, params.id = NULL, param.names = NULL,
  prior.id = NULL, chain = NULL, iter = NULL, adapt = NULL,
  adj.min = NULL, ar.target = NULL, jvar = NULL, n.knot = NULL,
  burnin = NULL)
}
\arguments{
\item{all}{params are the identically named variables in pda.mcmc / pda.emulator}
}
\value{
An updated settings list
}
\description{
Clean up a failed PDA run
}
\author{
Ryan Kelly
}

