% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pda.neff.R
\name{pda.neff.calc}
\alias{pda.neff.calc}
\title{Calculate N_eff}
\usage{
pda.neff.calc(inputs, recalculate = FALSE)
}
\arguments{
\item{inputs}{list}

\item{recalculate}{repeat neff calculation or not}
}
\value{
inputs list, updated inputs with n_eff
}
\description{
Autocorelation correction and efficient sample size calculation on latent process
}
\details{
What we're trying to do is to calculate the autocorrelation of the latent state, after attempting to "remove" the observation error.
The first step is thus to estimate the latent state using a simple 'process free' state-space model (e.g. random walk).
}
\author{
Istem Fer
}
