% Generated by roxygen2 (4.0.2): do not edit by hand
\name{pda.plot.params}
\alias{pda.plot.params}
\title{Plot PDA Parameter Diagnostics}
\usage{
pda.plot.params(settings, params.subset, prior.ind)
}
\arguments{
\item{all}{params are the identically named variables in pda.mcmc / pda.emulator}
}
\value{
Nothing. Plot is generated and saved to PDF.
}
\description{
Plot PDA Parameter Diagnostics
}
\author{
Ryan Kelly
}

