% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pda.utils.R
\name{pda.postprocess}
\alias{pda.postprocess}
\title{Postprocessing for PDA Results}
\usage{
pda.postprocess(settings, con, params, pname, prior, prior.ind, burnin = NULL)
}
\arguments{
\item{all}{params are the identically named variables in pda.mcmc / pda.emulator}
}
\value{
PEcAn settings list, updated with <params.id> pointing to the new params file.
}
\description{
Postprocessing for PDA Results
}
\author{
Ryan Kelly
}

