% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/pda.utils.R
\name{pda.settings}
\alias{pda.settings}
\title{Set PDA Settings}
\usage{
pda.settings(settings, params.id = NULL, param.names = NULL,
  prior.id = NULL, chain = NULL, iter = NULL, adapt = NULL,
  adj.min = NULL, ar.target = NULL, jvar = NULL, n.knot = NULL)
}
\arguments{
\item{all}{params are the identically named variables in pda.mcmc / pda.emulator}
}
\value{
An updated settings list
}
\description{
Set PDA Settings
}
\author{
Ryan Kelly
}

